# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_transport_http_v1.createTaskRequest:
    properties:
      goal_id:
        type: integer
      scheduled_date_time:
        type: string
      scheduled_end_date_time:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  internal_transport_http_v1.listTasksResponse:
    properties:
      task_data:
        items:
          $ref: '#/definitions/internal_transport_http_v1.taskData'
        type: array
      user_id:
        type: integer
    type: object
...
  internal_transport_http_v1.taskData:
    properties:
      created_at:
        type: string
      duration_minutes:
        type: integer
      goal_id:
        type: integer
      id:
        type: integer
      is_done:
        type: boolean
      reschedule_count:
        type: integer
      scheduled_date:
        type: string
      scheduled_time:
        type: string
      title:
        type: string
    type: object
  internal_transport_http_v1.taskResponse:
    properties:
      created_at:
        type: string
      duration_minutes:
        type: integer
      goal_id:
        type: integer
      id:
        type: integer
      is_done:
        type: boolean
      reschedule_count:
        type: integer
      scheduled_date:
        type: string
      scheduled_time:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  internal_transport_http_v1.updateTaskRequest:
    properties:
      goal_id:
        type: integer
      is_done:
        type: boolean
      scheduled_date_time:
        type: string
      scheduled_end_date_time:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API for Mile-Do, simple clone of TickTick.
  termsOfService: http://swagger.io/terms/
  title: Mile-Do API
  version: "1.0"
paths:
...
  /goals/{id}/tasks:
    get:
      consumes:
      - application/json
      description: get tasks by :goal_id
      parameters:
      - description: Goal ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.listTasksResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get tasks by :goal_id
      tags:
      - tasks
  /tasks/:
    get:
      consumes:
      - application/json
      description: get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.listTasksResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: create new task
      parameters:
      - description: Task Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http_v1.taskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: create new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: delete task by :id
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Task has been removed
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: delete task by :id
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: get task by :id
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.taskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get task by :id
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: update existing task by :id
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: New Task Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.taskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: update task by :id
      tags:
      - tasks
  /tasks/analyze:
    get:
      consumes:
      - application/json
      description: get count of completed tasks over total tasks for today
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get stats for today
      tags:
      - tasks
  /tasks/inbox:
    get:
      consumes:
      - application/json
      description: get tasks without date
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.listTasksResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get inbox tasks
      tags:
      - tasks
  /tasks/period:
    get:
      consumes:
      - application/json
      description: get tasks by period
      parameters:
      - description: tasks after specific date
        in: query
        name: after_date
        type: string
      - description: tasks before specific date
        in: query
        name: before_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.listTasksResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get tasks by period
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"